requirements.txt
streamlit
pandas
plotly
google-genai
pinecone
langchain-google-genai
langchain-pinecone
gspread
oauth2client
python-dotenv